Make component so it is understandable what happens. Also fix major bug where a ref was passed instead of destructuring... omg
